+++ extensions/gif-search/src/hooks/useFavorites.ts
@@ -5,18 +5,20 @@ import { ServiceName } from "../preferences";
 import { getAPIByServiceName } from "./useSearchAPI";
 
 import dedupe from "../lib/dedupe";
-import { clearFavorites, getFavorites } from "../lib/favorites";
+import { clearAllFavorites, clearFavorites, getAllFavorites, getFavorites } from "../lib/favorites";
 
 import type { IGif } from "../models/gif";
 
+type GifIds = Set<string>;
+
 export interface FavIdsState {
-  ids?: Set<string>;
+  ids?: Map<ServiceName, GifIds>;
   error?: Error;
 }
 
 export interface FavItemsState {
-  items?: IGif[];
-  error?: Error;
+  items?: Map<ServiceName, IGif[]>;
+  errors?: Error[];
 }
 
 const DEFAULT_RESULT_COUNT = 10;
@@ -38,7 +40,7 @@ export default function useFavorites({ offset = 0, limit = DEFAULT_RESULT_COUNT 
 
       try {
         const faveIds = await getFavorites(service);
-        setFavIds({ ids: faveIds });
+        setFavIds({ ids: new Map([[service, faveIds]]) });
       } catch (e) {
         console.error(e);
         const error = e as Error;
@@ -52,24 +54,49 @@ export default function useFavorites({ offset = 0, limit = DEFAULT_RESULT_COUNT 
     [setFavIds, setIsLoadingIds]
   );
 
+  const loadAllFavs = useCallback(
+    async function loadAllFavs() {
+      setIsLoadingIds(true);
+
+      try {
+        const allFavs = await getAllFavorites();
+        setFavIds({ ids: allFavs });
+      } catch (e) {
+        console.error(e);
+        const error = e as Error;
+
+        await clearAllFavorites();
+        setFavIds({ error });
+      } finally {
+        setIsLoadingIds(false);
+      }
+    },
+    [setFavIds, setIsLoadingFavs]
+  );
+
   const populate = useCallback(
-    async function populate(ids: Set<string>, service?: ServiceName) {
+    async function populate(ids: GifIds, service?: ServiceName) {
       setIsLoadingFavs(true);
 
       if (!service || !ids.size) {
-        setFavItems({ items: [] });
+        setFavItems({ items: new Map() });
         setIsLoadingFavs(false);
         return;
       }
 
       const api = await getAPIByServiceName(service);
+      if (!api) {
+        setFavItems({ items: new Map() });
+        setIsLoadingFavs(false);
+        return;
+      }
 
       let items: IGif[];
       try {
         items = dedupe(await api.gifs([...ids])).slice(0, limit);
         items?.forEach((item) => (item.is_fav = !!ids?.has(item.id.toString())));
 
-        setFavItems({ items });
+        items.length && setFavItems({ items: new Map([[service, items]]) });
       } catch (e) {
         const error = e as FetchError;
         if (e instanceof AbortError) {
@@ -80,7 +107,7 @@ export default function useFavorites({ offset = 0, limit = DEFAULT_RESULT_COUNT 
         } else {
           await clearFavorites(service);
         }
-        setFavItems({ error });
+        setFavItems({ errors: [error] });
       }
 
       setIsLoadingFavs(false);
@@ -88,5 +115,57 @@ export default function useFavorites({ offset = 0, limit = DEFAULT_RESULT_COUNT 
     [setIsLoadingFavs, setFavItems]
   );
 
-  return [favIds, favItems, isLoadingIds, isLoadingFavs, loadFavs, populate] as const;
+  const populateAll = useCallback(
+    async function populateAll(allIds: Map<ServiceName, GifIds>) {
+      setIsLoadingFavs(true);
+
+      if (!allIds.size) {
+        setFavItems({ items: new Map() });
+        setIsLoadingFavs(false);
+        return;
+      }
+
+      const allItems = new Map<ServiceName, IGif[]>();
+      const errors: Error[] = [];
+
+      for (const service of allIds.keys()) {
+        const api = await getAPIByServiceName(service);
+        if (!api) {
+          continue;
+        }
+
+        let items: IGif[];
+        try {
+          const ids = allIds.get(service) || new Set<string>();
+          items = dedupe(await api.gifs([...ids])).slice(0, limit);
+          items?.forEach((item) => (item.is_fav = !!ids?.has(item.id.toString())));
+
+          items.length && allItems.set(service, items);
+        } catch (e) {
+          const error = e as FetchError;
+          if (e instanceof AbortError) {
+            return;
+          } else if (error.message.toLowerCase().includes("invalid authentication credentials")) {
+            error.message = "Invalid credentials, please try again.";
+            await getAPIByServiceName(service, true);
+          } else {
+            await clearFavorites(service);
+          }
+          errors.push(error);
+          break;
+        }
+      }
+
+      if (errors.length) {
+        setFavItems({ errors });
+      } else {
+        setFavItems({ items: allItems });
+      }
+
+      setIsLoadingFavs(false);
+    },
+    [setIsLoadingFavs, setFavItems]
+  );
+
+  return [favIds, favItems, isLoadingIds, isLoadingFavs, loadFavs, populate, loadAllFavs, populateAll] as const;
 }