+++ extensions/gif-search/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add section dividers to GIF Details Actions
 - Add custom Empty State
 - Add preference for max number of search results (defaults to 10)
+- Add Favorites view so you can see and search all your favorite gifs across all services
 
 ## [Remove static GIF preview warning 🎉] - 2022-04-06
 - Removes the warning about GIFs no longer animating, Raycast v1.32.0 supports GIFs in Details